Revert "refactor: use fixture filename in todo example for footer and header"
This reverts commit acf7fa0bef5d41ba8b96fd60e0f51116bfa78572.
