+++ examples/todo/src/components/Header.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { useTodoContext } from '../TodoContext.js';
+import { useTodoContext } from './TodoContext.js';
 
 export function Header() {
   const { todos, setTodos } = useTodoContext();
